chore: move the move bitboards to a different module
