+++ games/src/ataxx/position.rs
@@ -30,6 +30,8 @@ use crate::ataxx::{
 use crate::MoveList;
 use crate::MoveStore;
 
+use super::moves;
+
 /// Position represents the snapshot of an Ataxx Board, the state of the an
 /// ataxx game at a single point in time. It also provides all of the methods
 /// necessary to manipulate such a snapshot.
@@ -264,7 +266,7 @@ impl Position {
         let stm_pieces = self.color_bb(stm);
         let xtm_pieces = self.color_bb(!stm);
 
-        let captured = BitBoard::single(m.target()) & xtm_pieces;
+        let captured = moves::single(m.target()) & xtm_pieces;
         let from_to = BitBoard::from(m.target()) | BitBoard::from(m.source());
 
         // Move the captured pieces from xtm to stm.
@@ -349,7 +351,7 @@ impl Position {
         for piece in stm {
             // There may be multiple jump moves to a single Square, so they need to be
             // verified (& allowed) and serialized into the movelist immediately.
-            let double = BitBoard::double(piece) & allowed;
+            let double = moves::double(piece) & allowed;
             for target in double {
                 movelist.push(Move::new(piece, target));
             }
@@ -393,7 +395,7 @@ impl Position {
         for piece in stm {
             // There may be multiple jump moves to a single Square, so they need to be
             // verified (& allowed) and counted into the Position total immediately.
-            let double = BitBoard::double(piece) & allowed;
+            let double = moves::double(piece) & allowed;
             moves += double.cardinality();
         }
